Refactor charts' usage
